# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Kols API",
        "contact": {
            "name": "Ken"
        },
        "version": "1.0"
    },
    "host": "api.kols-ammedia.local",
    "basePath": "/v1",
    "schemes": [
        "http"
    ],
    "paths": {
        "/facebook-analytics": {
            "get": {
                "tags": [
                    "facebook analytics"
                ],
                "summary": "Facebook Overview API",
                "description": "get facebook data for table overview",
                "operationId": "getListFacebookAnalytics",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
        "/facebook-analytics/{profile_id}": {
            "get": {
                "tags": [
                    "facebook analytics"
                ],
                "summary": "The Detail of Page/Profile",
                "description": "get detail of facebook profile via profile id",
                "operationId": "getFacebookAnalyticsByProfileID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "profile_id",
                        "in": "path",
                        "description": "profile id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
        "/facebook-analytics/{profile_id}/most-engaging-posts?last_days={last_days}": {
            "get": {
                "tags": [
                    "facebook analytics"
                ],
                "summary": "Most engaging posts via profile id",
                "description": "get most engaging posts post via profile id",
                "operationId": "getFacebookMostEngagingPostsByProfileID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "profile_id",
                        "in": "path",
                        "description": "profile id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "last_days",
                        "in": "path",
                        "description": "Number of days ago",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
        "/facebook-analytics/{profile_id}/growth-total-fans?last_days={last_days}": {
            "get": {
                "tags": [
                    "facebook analytics"
                ],
                "summary": "Fan overview: Growth of total fans",
                "description": "analytics growth of total fans in days via profile_id and number of days ago",
                "operationId": "getFacebookGrowthOfTotalFan",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "profile_id",
                        "in": "path",
                        "description": "profile id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "last_days",
                        "in": "path",
                        "description": "Number of days ago",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
        "/facebook-analytics/{profile_id}/analytics-posts-per-days?last_days={last_days}": {
            "get": {
                "tags": [
                    "facebook analytics"
                ],
                "summary": "Content Overview: Number of Page Posts",
                "description": "analytics total posts in days via profile_id and number of days ago",
                "operationId": "getFacebookAnalyticsByProfileID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "profile_id",
                        "in": "path",
                        "description": "profile id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "last_days",
                        "in": "path",
                        "description": "Number of days ago",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
        "/facebook-analytics/{profile_id}/distribution-page-post-types?last_days={last_days}": {
            "get": {
                "tags": [
                    "facebook analytics"
                ],
                "summary": "Distribution of Page Post Types",
                "description": "analytics distribution of page post type via profile_id and number of days ago",
                "operationId": "getFacebookAnalyticsByProfileID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "profile_id",
                        "in": "path",
                        "description": "profile id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "last_days",
                        "in": "path",
                        "description": "Number of days ago",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
        "/facebook-analytics/{profile_id}/number-of-fan-posts?last_days={last_days}": {
            "get": {
                "tags": [
                    "facebook analytics"
                ],
                "summary": "Number of fan post",
                "description": "analytics number of fan posts via profile_id and number of days ago",
                "operationId": "analyticsNumberOfFanPost",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "profile_id",
                        "in": "path",
                        "description": "profile id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "last_days",
                        "in": "path",
                        "description": "Number of days ago",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
        "/instagram-analytics/profiles": {
            "get": {
                "tags": [
                    "instagram analytics"
                ],
                "summary": "Instagram Overview: List Instagram Profiles",
                "description": "get list instagram profiles",
                "operationId": "getListInstagramProfiles",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
        "/instagram-analytics/{username}/media": {
            "get": {
                "tags": [
                    "instagram analytics"
                ],
                "summary": "Instagram Overview: List Media Instagram",
                "description": "get all media via username",
                "operationId": "getInstagramMediaViaUsername",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "username",
                        "in": "path",
                        "description": "instagram username",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        },
        "/youtube-analytics/channels": {
            "get": {
                "tags": [
                    "youtube analytics"
                ],
                "summary": "Youtube Overview: List Youtube channels",
                "description": "get list youtube channels",
                "operationId": "getListYoutubeChannels",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "Internal Error"
                    }
                }
            }
        }
    },
    "definitions": {}
}